Try clean-up code organizations (wip)
